+++ ui-admin/src/study/participants/participantList/search/BasicSearch.tsx
@@ -1,7 +1,8 @@
-import React, { useState } from 'react'
+import React, { useCallback, useEffect, useState } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faSearch } from '@fortawesome/free-solid-svg-icons'
-import { ParticipantSearchState } from './ParticipantSearch'
+import { ParticipantSearchState } from 'util/participantSearchUtils'
+import { debounce } from 'lodash'
 
 /**
  * renders and manages updates to a string search facet
@@ -11,17 +12,35 @@ const BasicSearch = ({ searchState, updateSearchState }: {
   updateSearchState: (field: keyof ParticipantSearchState, value: unknown) => void
 }) => {
   const [search, setSearch] = useState(searchState.basicSearch)
+
+  useEffect(() => {
+    setSearch(searchState.basicSearch)
+  }, [searchState.basicSearch])
+
+  // downloading all the participant data is expensive, so debounce the searchbar
+  const debouncedUpdate = useCallback(
+    debounce(value => {
+      updateSearchState('basicSearch', value)
+    }, 500), []
+  )
+
   return <form className="rounded-5" onSubmit={e => {
     e.preventDefault()
-    updateSearchState('basicSearch', search)
+    debouncedUpdate(search)
   }} style={{ border: '1px solid #bbb', backgroundColor: '#fff', padding: '0.25em 0.75em 0em' }}>
     <button type="submit" title="submit search" className="btn btn-secondary">
       <FontAwesomeIcon icon={faSearch}/>
     </button>
-    <input type="text" value={search} size={40}
+    <input
+      type="text"
+      value={search}
+      size={40}
       style={{ border: 'none', outline: 'none' }}
       placeholder={'Search by name, email, or shortcode'}
-      onChange={e => setSearch(e.target.value)}/>
+      onChange={e => {
+        setSearch(e.target.value)
+        debouncedUpdate(e.target.value)
+      }}/>
   </form>
 }
