+++ ui-admin/src/study/participants/participantList/search/SearchCriteriaView.test.tsx
@@ -3,19 +3,16 @@ import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import SearchCriteriaView from './SearchCriteriaView'
 import { setupRouterTest } from '@juniper/ui-core'
+import { defaultParticipantSearchState } from '../../../../util/participantSearchUtils'
 
 describe('SearchCriteriaView', () => {
   test('shows and deletes search criteria', async () => {
     const mockUpdateSearchStateFn = jest.fn()
     const { RoutedComponent } = setupRouterTest(
       <SearchCriteriaView searchState={{
-        basicSearch: '',
-        minAge: undefined,
-        maxAge: undefined,
+        ...defaultParticipantSearchState,
         sexAtBirth: ['F'],
-        tasks: [{ task: 'consent', status: 'complete' }],
-        latestKitStatus: [],
-        custom: ''
+        tasks: [{ task: 'consent', status: 'complete' }]
       }} updateSearchState={mockUpdateSearchStateFn}/>)
     render(RoutedComponent)
